beta progress bar development 8th june 18
This file is for use of showing progress bar with dynamically changing its value according to value provided to it. It uses bootstrap plugin for progress bar. It shows how to create a jQuery plugin using a Javascript object, its prototype and communicate between DOM element through the jQuery data-api